chore(lang-css): restructure tests folders
Split the tests folders into subfolders to separate (integration) tests,
benchmarks, and utilities.
